backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "manuscript-layout-api"
version = "1.0.0"
description = "FastAPI backend for medieval manuscript layout analysis"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.18",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.0.0",
    "ultralytics>=8.0.0",
    "numpy<2.0",
    "pillow>=10.0.0",
    "matplotlib>=3.6.0",
    "shapely>=2.0.0",
    "rtree>=1.0.0",
    "langdetect>=1.0.9",
    "openai>=1.51.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]
